pyproject.toml
[build-system]
requires = ["setuptools>=67.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "note_gen"
version = "0.2.0"
description = "Note Generation API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "note_gen",
    "note_gen.routers",
    "note_gen.models",
    "note_gen.core"
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
pythonpath = [".", "src"]

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "src"
]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
namespace_packages = true
mypy_path = ["src"]
exclude = ["tools", "setup.py"]

[[tool.mypy.overrides]]
module = ["src.note_gen.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = ["html5lib.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["setuptools.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["motor.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pymongo.*"]
ignore_missing_imports = true

[tool.poetry.dependencies]
pytest-asyncio = "^0.23.0"
pytest-timeout = "^2.1.0"
